kot5/requirements.txt
t5==0.6.4
mesh-tensorflow==0.1.17
tensorflow-datasets==4.1.0